% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_kinbank.R
\name{download_kinbank}
\alias{download_kinbank}
\title{Downloading a version of Kinbank}
\usage{
download_kinbank(version = "latest", destination = ".")
}
\arguments{
\item{version}{the version of kinbank to download. Currently you can only download the latest version}

\item{destination}{the destination for the dataset to be saved to}
}
\value{
This function returns folder containing the raw and cldf versino of kinbank.
}
\description{
Downloading a version of Kinbank
}
\examples{
## not tested
# download_kinbank()
}
