% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinterms.R
\name{get_dist}
\alias{get_dist}
\title{Calculate distances within a language}
\usage{
get_dist(x, y, max_char, method = "osa")
}
\arguments{
\item{x}{a kinterm}

\item{y}{a kinterm to compare to}

\item{max_char}{the longest kinterm in the langauge}

\item{method}{the comparison type to be used (binary or string distance)}
}
\value{
a distance between two kin terms.
}
\description{
Only for internal use.
}
