% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinterms.R
\name{get_structural_vectors}
\alias{get_structural_vectors}
\title{Extract Kinterms from Kinbank into a matrix}
\usage{
get_structural_vectors(
  kin_types,
  duplicates,
  languages = NULL,
  method = "binary"
)
}
\arguments{
\item{kin_types}{a character list of kintypes to extract. See kin_types() for a list of possibilities.}

\item{duplicates}{In some cases, languages have more than one kinterm per category.}

\item{languages}{a vector of Language IDs for which to calculate structural vectors for.}

\item{method}{the method for which to calculate the structural vectory. Options: binary (more to come).}
}
\value{
a matrix with columns of kin types, and rows with languages. Each cell contains a kinterm.
}
\description{
This function will output a matrix of kinterms, where columns are kin types, and rows are languages.
Some languages have more than one kin term per kin type. Duplicates is an argument that allows us how to decide how to deal with these.
}
