% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinterms.R
\name{get_vector}
\alias{get_vector}
\title{Extrapolate the structural properties of a kin term subset.}
\usage{
get_vector(kin_data, method = "binary")
}
\arguments{
\item{kin_data}{The list of kinterms and kindata}

\item{method}{How should comparisons be calculated? Usually binary.}
}
\description{
This is an internally used function.
}
