% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kin_terms}
\alias{kin_terms}
\title{Kin terms}
\format{
A data frame with 210903 rows and 13 variables:
\describe{
  \item{\code{ID}}{character A unique identifier.}
  \item{\code{Local_ID}}{logical <To be removed>}
  \item{\code{Language_ID}}{character ID from the langauges datatable.}
  \item{\code{Parameter_ID}}{character ID from the parameters datatable.}
  \item{\code{Value}}{character Contains the original transcription of the data. May include multiple terms per row (e.g. "brother; bro). Please verify.}
  \item{\code{Form}}{character Contains a cleaned transcription of the data, splitting data into one term per row. Please check the data has been correctly parsed.}
  \item{\code{Segments}}{logical <ignore>}
  \item{\code{Comment}}{character Comments on a kinterm, usually offering more detail or considerations.}
  \item{\code{Source}}{character Bibtex code for the source of the kinterm.}
  \item{\code{Cognacy}}{logical <future column for cognates>}
  \item{\code{Loan}}{logical <future column for labelling loans>}
  \item{\code{Graphemes}}{logical <future column for labelling graphemes>}
  \item{\code{Profile}}{logical =ignore>}
}
}
\source{
\url{http://www.kinbank.net}
}
\usage{
kin_terms
}
\description{
This contains the forms, or kinterms, for each language. This file links to the languages file by the Languages_ID column, parameters file by the Parameters_ID column, and the sources file by the sources_bibtex column.
}
\keyword{datasets}
