% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Kin types available within Kinbank}
\format{
A data frame with 943 rows and 7 variables:
\describe{
  \item{\code{ID}}{character The unique identifier for the parameter. Often the same as Parameter, but without special characters.}
  \item{\code{Name}}{character A written description of ID}
  \item{\code{Concepticon_ID}}{integer A numeric identifier to align with Concepticon (https://concepticon.clld.org/)}
  \item{\code{Concepticon_Gloss}}{character A description of the concepticon identifier}
  \item{\code{Parameter}}{character The parameter used when coding Kinbank. Usually, this is built from primitive kinship abbreviations, such as: B (brother), Z (sister), F (father), M (mother), S (son), D (daughter), C (child), H (husband), W (wife). As well as using modifiers such as e (elder), y (younger), m (male speaker), f (female speaker)}
  \item{\code{Group}}{character Broader categorisation of the parameter (e.g. The parameter B (brother) is categorised more broadly as Sibling)}
  \item{\code{Dataset}}{character Describes which Kinbank collaborator project collected the data}
}
}
\source{
\url{http://www.kinbank.net/}
}
\usage{
parameters
}
\description{
This file contains descriptions of each kin type used in the dataset. Some kin types are used only once, others are used often. For the template set of kinterms, see the Kinbank publication.
}
\keyword{datasets}
