% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revert_vector.R
\name{revert_vector}
\alias{revert_vector}
\title{Revert Structural vectors}
\usage{
revert_vector(v, nmes = "none")
}
\arguments{
\item{v}{the binarised structural vector from get_structural_vector()}

\item{nmes}{the names of parameters used in get_structural_vector()}
}
\value{
a character vector representing the original kin terminology set, with letters in place of kinterms
}
\description{
Revert Structural vectors
}
