% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{show_kin_types}
\alias{show_kin_types}
\title{List all parameters used in Kinbank.}
\usage{
show_kin_types()
}
\value{
A data.frame of all kin types in the Kinbank dataset show the parameter and written description.
}
\description{
List all parameters used in Kinbank.
}
\examples{
show_kin_types()
}
