% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{show_languages}
\alias{show_languages}
\title{List all languages in Kinbank}
\usage{
show_languages()
}
\value{
a character vector of all languages in Kinbank
}
\description{
List all languages in Kinbank
}
\examples{
show_languages()
}
